+++ build.sh
@@ -1,0 +1,6 @@
+
+#!/bin/sh
+
+make ci-generate ci-build ci-test ci-deploy
+GO111MODULE="on" make ci-soups
+echo "build done"