+++ local/env.build.template.sh
@@ -1,0 +1,16 @@
+export DOCKER_REGISTRY=docker.ci.diabeloop.eu
+export ARTIFACT_NODE_VERSION=0.0.0
+export TRAVIS_NODE_VERSION=0.0.0
+
+export TRAVIS_BRANCH=a-feature
+export TRAVIS_PULL_REQUEST=A-PR
+export TRAVIS_TAG=a.tag
+export TRAVIS_REPO_SLUG=mdblp/platform
+export TRAVIS_PULL_REQUEST_SHA=123456789
+export TRAVIS_GO_VERSION=1.11.4
+export SECURITY_SCAN=true
+export OPS_DOCKER_REGISTRY=registry.coreye.fr/diabeloop/artifacts/diabeloop_docker
+export OPS_DOCKER_USERNAME=user
+export OPS_DOCKER_PASSWORD="password"
+export DOCKER_USERNAME=user
+export DOCKER_PASSWORD="password"